migrate all justice commands to their own files/commands
todo: history
